# quizonline-frontend/src/app/api/generated/model/custom-user-read.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LanguageEnumDto } from './language-enum';


export interface CustomUserReadDto { 
    readonly id: number;
    /**
     * Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.
     */
    readonly username: string;
    email?: string;
    first_name?: string;
    last_name?: string;
    language?: LanguageEnumDto;
    readonly email_confirmed: boolean;
    readonly password_change_required: boolean;
    /**
     * Designates that this user has all permissions without explicitly assigning them.
     */
    readonly is_superuser: boolean;
    /**
     * Designates whether the user can log into this admin site.
     */
    readonly is_staff: boolean;
    /**
     * Designates whether this user should be treated as active. Unselect this instead of deleting accounts.
     */
    readonly is_active: boolean;
    readonly nb_domain_max: number;
    current_domain?: number | null;
    readonly current_domain_title: string;
    readonly owned_domain_ids: Array<number>;
    readonly managed_domain_ids: Array<number>;
}