# quizonline-server/openapi.yaml
openapi: 3.0.3
info:
  title: QuizOnline API
  version: 1.0.0
paths:
...
  /api/user/{user_id}/quizzes/:
    get:
      operationId: user_quiz_list
      description: |-
        Retourne la liste des quiz liés à un utilisateur.
        Accès: soi-même ou staff/superuser.
      summary: Lister les quizzes d’un utilisateur
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - in: path
        name: user_id
        schema:
          type: integer
        description: ID utilisateur.
        required: true
      tags:
      - User
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedQuizSimpleList'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Not found
...
  /api/user/me/:
    get:
      operationId: user_me_retrieve
      description: Retourne le profil de l'utilisateur authentifié.
      summary: Récupérer et mettre à jour mon profil
      tags:
      - User
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomUserRead'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Validation error
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Forbidden
    patch:
      operationId: user_me_partial_update
      description: Retourne le profil de l'utilisateur authentifié.
      summary: Récupérer et mettre à jour mon profil
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCustomUserProfileUpdateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCustomUserProfileUpdateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCustomUserProfileUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomUserRead'
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Validation error
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Forbidden
  /api/user/me/current-domain/:
    post:
      operationId: user_me_current_domain_create
      description: Met à jour `current_domain` sur l'utilisateur authentifié, puis
        renvoie le profil mis à jour.
      summary: Définir mon domaine courant
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetCurrentDomainRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SetCurrentDomainRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SetCurrentDomainRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomUserRead'
          description: ''
        '400':
          description: Validation error
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
          description: Forbidden
...
components:
  schemas:
...
    CustomUserRead:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        username:
          type: string
          readOnly: true
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        language:
          $ref: '#/components/schemas/LanguageEnum'
        email_confirmed:
          type: boolean
          readOnly: true
        password_change_required:
          type: boolean
          readOnly: true
        is_superuser:
          type: boolean
          readOnly: true
          title: Superuser status
          description: Designates that this user has all permissions without explicitly
            assigning them.
        is_staff:
          type: boolean
          readOnly: true
          title: Staff status
          description: Designates whether the user can log into this admin site.
        is_active:
          type: boolean
          readOnly: true
          title: Active
          description: Designates whether this user should be treated as active. Unselect
            this instead of deleting accounts.
        nb_domain_max:
          type: integer
          readOnly: true
        current_domain:
          type: integer
          nullable: true
        current_domain_title:
          type: string
          readOnly: true
        owned_domain_ids:
          type: array
          items:
            type: integer
          readOnly: true
        managed_domain_ids:
          type: array
          items:
            type: integer
          readOnly: true
      required:
      - current_domain_title
      - email_confirmed
      - id
      - is_active
      - is_staff
      - is_superuser
      - managed_domain_ids
      - nb_domain_max
      - owned_domain_ids
      - password_change_required
      - username
...
    ErrorDetail:
      type: object
      properties:
        detail:
          type: string
...
    LanguageEnum:
      enum:
      - en
      - fr
      - nl
      - it
      - es
      type: string
      description: |-
        * `en` - English
        * `fr` - French
        * `nl` - Dutch
        * `it` - Italy
        * `es` - Spain
...
    PaginatedQuizSimpleList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/QuizSimple'
...
    PatchedCustomUserProfileUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          title: Email address
          maxLength: 254
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        language:
          $ref: '#/components/schemas/LanguageEnum'
        managed_domain_ids:
          type: array
          items:
            type: integer
            minimum: 1
...
    QuizSimple:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          readOnly: true
      required:
      - id
      - title
...
    SetCurrentDomainRequest:
      type: object
      properties:
        domain_id:
          type: integer
          nullable: true
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT